# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "operationId": "getCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update current user profile",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/me/change-password": {
      "post": {
        "operationId": "changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully"
          },
          "400": {
            "description": "Invalid current password"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Change user password",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/rooms": {
      "post": {
        "operationId": "createRoom",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Rooms"
        ]
      },
      "get": {
        "operationId": "getRooms",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "landlordId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "buildingId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Room"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Rooms"
        ]
      }
    },
    "/api/v1/rooms/{id}": {
      "get": {
        "operationId": "getRoom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Rooms"
        ]
      },
      "patch": {
        "operationId": "updateRoom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Rooms"
        ]
      },
      "delete": {
        "operationId": "deleteRoom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Rooms"
        ]
      }
    },
...
  },
  "info": {
    "title": "TacoHouse API",
    "description": "A comprehensive rental room management system API for managing buildings, rooms, tenants, billing, and payments.\n\n## Response Format\nAll responses follow a standardized format:\n- **status**: HTTP status code\n- **message**: Human-readable message\n- **data**: Response data (object or array)\n- **pagination**: (Optional) Pagination metadata for list endpoints",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3005",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
...
      "UserRole": {
        "type": "string",
        "enum": [
          "ADMIN",
          "LANDLORD",
          "TENANT"
        ]
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "avatar": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "occupation": {
            "type": "string"
          },
          "workplace": {
            "type": "string"
          },
          "idCardFrontPhoto": {
            "type": "string",
            "nullable": true
          },
          "idCardBackPhoto": {
            "type": "string",
            "nullable": true
          },
          "portraitPhoto": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "firstName",
          "lastName",
          "phone",
          "avatar",
          "dateOfBirth",
          "occupation",
          "workplace",
          "idCardFrontPhoto",
          "idCardBackPhoto",
          "portraitPhoto",
          "createdAt",
          "updatedAt"
        ]
      },
      "Session": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "ipAddress": {
            "type": "string",
            "nullable": true
          },
          "userAgent": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "token",
          "expiresAt",
          "ipAddress",
          "userAgent",
          "createdAt",
          "updatedAt"
        ]
      },
      "Account": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "providerId": {
            "type": "string"
          },
          "accountId": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "accessToken": {
            "type": "string",
            "nullable": true
          },
          "refreshToken": {
            "type": "string",
            "nullable": true
          },
          "idToken": {
            "type": "string",
            "nullable": true
          },
          "accessTokenExpiresAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "refreshTokenExpiresAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "scope": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "providerId",
          "accountId",
          "password",
          "accessToken",
          "refreshToken",
          "idToken",
          "accessTokenExpiresAt",
          "refreshTokenExpiresAt",
          "scope",
          "createdAt",
          "updatedAt"
        ]
      },
      "RoomType": {
        "type": "string",
        "enum": [
          "FULL_RIGHTS",
          "PARTIAL_RIGHTS"
        ]
      },
...
      "RentalStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "NOTICE_GIVEN",
          "TERMINATED"
        ]
      },
      "Rental": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "tenantRole": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "roomId": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "noticeDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "monthlyRent": {
            "type": "string",
            "format": "Decimal.js"
          },
          "depositPaid": {
            "type": "string",
            "format": "Decimal.js"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RentalStatus"
              }
            ]
          },
          "contractImages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "tenantId",
          "tenantRole",
          "roomId",
          "startDate",
          "endDate",
          "noticeDate",
          "monthlyRent",
          "depositPaid",
          "status",
          "contractImages",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "PaymentConfirmation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "paymentId": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "tenantRole": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "tenantConfirmed": {
            "type": "boolean"
          },
          "tenantConfirmedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "landlordConfirmed": {
            "type": "boolean"
          },
          "landlordConfirmedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "proofImages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "paymentId",
          "tenantId",
          "tenantRole",
          "tenantConfirmed",
          "tenantConfirmedAt",
          "landlordConfirmed",
          "landlordConfirmedAt",
          "proofImages",
          "notes",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "MaintenanceRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "tenantRole": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "roomId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priority": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PriorityType"
              }
            ]
          },
          "category": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MaintenanceCategory"
              }
            ]
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MaintenanceStatus"
              }
            ]
          },
          "completedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "completionNote": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "tenantId",
          "tenantRole",
          "roomId",
          "title",
          "description",
          "priority",
          "category",
          "images",
          "status",
          "completedAt",
          "completionNote",
          "createdAt",
          "updatedAt"
        ]
      },
      "Room": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "buildingId": {
            "type": "string"
          },
          "number": {
            "type": "string"
          },
          "area": {
            "type": "string",
            "format": "Decimal.js"
          },
          "monthlyRent": {
            "type": "string",
            "format": "Decimal.js"
          },
          "deposit": {
            "type": "string",
            "format": "Decimal.js"
          },
          "maxTenants": {
            "type": "integer",
            "format": "int32"
          },
          "roomType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RoomType"
              }
            ]
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RoomStatus"
              }
            ]
          },
          "availableFrom": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "rentals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rental"
            }
          },
          "equipment": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoomEquipment"
            }
          },
          "utilityRecords": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UtilityRecord"
            }
          },
          "bills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Bill"
            }
          },
          "maintenanceRequests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MaintenanceRequest"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "buildingId",
          "number",
          "area",
          "monthlyRent",
          "deposit",
          "maxTenants",
          "roomType",
          "description",
          "images",
          "status",
          "availableFrom",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "senderId": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "messageType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MessageType"
              }
            ]
          },
          "attachmentUrl": {
            "type": "string",
            "nullable": true
          },
          "attachmentName": {
            "type": "string",
            "nullable": true
          },
          "attachmentSize": {
            "type": "integer",
            "nullable": true,
            "format": "int32"
          },
          "chatGroupId": {
            "type": "string",
            "nullable": true
          },
          "recipientId": {
            "type": "string",
            "nullable": true
          },
          "readAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "senderId",
          "content",
          "messageType",
          "attachmentUrl",
          "attachmentName",
          "attachmentSize",
          "chatGroupId",
          "recipientId",
          "readAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "ChatGroupMember": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "chatGroupId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "chatGroupId",
          "userId",
          "joinedAt"
        ]
      },
...
      "Building": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "landlordId": {
            "type": "string"
          },
          "landlordRole": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "billingDate": {
            "type": "integer",
            "nullable": true,
            "format": "int32"
          },
          "electricityRate": {
            "type": "string",
            "format": "Decimal.js"
          },
          "waterRate": {
            "type": "string",
            "format": "Decimal.js"
          },
          "gasRate": {
            "type": "string",
            "format": "Decimal.js"
          },
          "managementFee": {
            "type": "string",
            "format": "Decimal.js"
          },
          "cleaningFeePerPerson": {
            "type": "string",
            "format": "Decimal.js"
          },
          "lightingFee": {
            "type": "string",
            "format": "Decimal.js"
          },
          "rooms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Room"
            }
          },
          "chatGroup": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ChatGroup"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "landlordId",
          "landlordRole",
          "name",
          "address",
          "description",
          "billingDate",
          "electricityRate",
          "waterRate",
          "gasRate",
          "managementFee",
          "cleaningFeePerPerson",
          "lightingFee",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "Notification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/NotificationType"
              }
            ]
          },
          "isRead": {
            "type": "boolean"
          },
          "readAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "relatedId": {
            "type": "string",
            "nullable": true
          },
          "relatedType": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedEntityType"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "title",
          "message",
          "type",
          "isRead",
          "readAt",
          "relatedId",
          "relatedType",
          "createdAt",
          "updatedAt"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "profile": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserProfile"
              }
            ]
          },
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Session"
            }
          },
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Account"
            }
          },
          "buildings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Building"
            }
          },
          "rentals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rental"
            }
          },
          "maintenanceRequests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MaintenanceRequest"
            }
          },
          "paymentConfirmations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentConfirmation"
            }
          },
          "sentMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "receivedMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "notifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notification"
            }
          },
          "chatGroupMember": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatGroupMember"
            }
          },
          "deletedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "isActive",
          "deletedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateUserProfileDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "minLength": 6,
            "maxLength": 50
          },
          "confirmPassword": {
            "type": "string",
            "minLength": 6,
            "maxLength": 50
          }
        },
        "required": [
          "currentPassword",
          "confirmPassword"
        ]
      },
...
      "CreateRoomDto": {
        "type": "object",
        "properties": {
          "number": {
            "type": "string"
          },
          "buildingId": {
            "type": "string"
          },
          "area": {
            "type": "number",
            "minimum": 1
          },
          "monthlyRent": {
            "type": "number",
            "minimum": 1
          },
          "deposit": {
            "type": "number",
            "default": 0,
            "minimum": 1
          },
          "maxTenants": {
            "type": "number",
            "default": 1,
            "minimum": 1
          },
          "roomType": {
            "type": "string",
            "enum": [
              "FULL_RIGHTS",
              "PARTIAL_RIGHTS"
            ]
          },
          "description": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "OCCUPIED",
              "PENDING_CHECKOUT",
              "MAINTENANCE"
            ]
          },
          "availableFrom": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "number",
          "buildingId",
          "area",
          "monthlyRent",
          "deposit",
          "maxTenants",
          "roomType",
          "status"
        ]
      },
      "UpdateRoomDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}